refactor: Extract the Actor Crafting App Header into its own component
- Create ActorCraftingAppHeader.svelte
- Move actor selection logic into the new component and out of the parent
- Share state through event propagation
- Reload recipes and components on sourceActorChanged event
