+++ src/applications/actorCraftingApp/ActorCraftingApp.svelte
@@ -1,14 +1,12 @@
 <!-- ActorCraftingApp.svelte -->
 <script lang="ts">
 
-    import {AppBar, AppRail, AppRailTile, AppShell, Avatar, ListBox, ListBoxItem} from "@skeletonlabs/skeleton";
+    import {AppBar, Avatar} from "@skeletonlabs/skeleton";
     import type {LocalizationService} from "../common/LocalizationService";
     import type {FabricateAPI} from "../../scripts/api/FabricateAPI";
-    import {ActorCraftingAppViewType} from "./ActorCraftingAppViewType";
     import type {ActorDetails} from "./ActorDetails";
-    import {NoActorDetails} from "./ActorDetails";
+    import {PendingActorDetails} from "./ActorDetails";
     import {onMount} from "svelte";
-    import {DefaultActorDetails} from "./ActorDetails.js";
     import type {CraftingAssessment} from "./CraftingAssessment";
     import truncate from "../common/Truncate";
     import Properties from "../../scripts/Properties";
@@ -16,7 +14,7 @@
     import {NoSalvageProcess, type SalvageProcess} from "./SalvageProcess";
     import type {SalvageAssessment} from "./SalvageAssessment";
     import {DefaultSalvageProcess} from "./SalvageProcess.js";
-    import {clickOutside} from "../common/ClickOutside";
+    import ActorCraftingAppHeader from "./ActorCraftingAppHeader.svelte";
 
     /*
      * ===========================================================================
@@ -24,7 +22,6 @@
      * ===========================================================================
      */
 
-    export let view: ActorCraftingAppViewType = ActorCraftingAppViewType.SALVAGING;
     export let localization: LocalizationService;
     export let fabricateAPI: FabricateAPI;
     export let sourceActorId: string;
@@ -38,25 +35,10 @@
      * ===========================================================================
      */
 
-    let sourceActorDetails: ActorDetails = new NoActorDetails();
-    let targetActorDetails: ActorDetails = new NoActorDetails();
-    let showSourceActorSelection: boolean = false;
+    let targetActorDetails: ActorDetails = new PendingActorDetails({id: targetActorId})
+    let sourceActorDetails: ActorDetails = new PendingActorDetails({id: sourceActorId ?? targetActorId});
     let craftingProcess: CraftingProcess = new NoCraftingProcess();
     let salvageProcess: SalvageProcess = new NoSalvageProcess();
-
-    /*
-     * ===========================================================================
-     * Recipe crafting assessments search
-     * ===========================================================================
-     */
-    let sourceActorName: string = "";
-    let availableSourceActors: ActorDetails[] = [];
-    $: filteredSourceActors = availableSourceActors.filter(actor => {
-        if (!sourceActorName) {
-            return true;
-        }
-        return actor.name.search(new RegExp(sourceActorName, "i")) >= 0;
-    });
 
     /*
      * ===========================================================================
@@ -115,65 +97,6 @@
      * ===========================================================================
      */
 
-    async function loadActorDetails() {
-        const sourceActor = await game.actors.get(sourceActorId);
-        if (!sourceActor) {
-            throw new Error(`Fabricate | ActorCraftingApp: Unable to find actor with id ${sourceActorId}`);
-        }
-        sourceActorDetails = new DefaultActorDetails({
-            id: sourceActorId,
-            name: sourceActor.name,
-            avatarUrl: sourceActor.img,
-            initials: DefaultActorDetails.getInitialsFromName(sourceActor.name)
-        });
-        if (targetActorId === sourceActorId) {
-            targetActorDetails = sourceActorDetails;
-            return;
-        }
-        const targetActor = await game.actors.get(targetActorId);
-        if (!targetActor) {
-            throw new Error(`Fabricate | ActorCraftingApp: Unable to find actor with id ${targetActorId}`);
-        }
-        targetActorDetails = new DefaultActorDetails({
-            id: targetActorId,
-            name: targetActor.name,
-            avatarUrl: targetActor.img,
-            initials: DefaultActorDetails.getInitialsFromName(targetActor.name)
-        });
-    }
-
-    async function setSourceActor(value: ActorDetails) {
-        sourceActorDetails = value;
-        showSourceActorSelection = false;
-        await prepareRecipes();
-        await prepareComponents();
-    }
-
-    async function clearSourceActor() {
-        sourceActorDetails = targetActorDetails;
-        await prepareRecipes();
-        await prepareComponents();
-    }
-
-    function toggleSourceActorSelectionMenu() {
-        if (showSourceActorSelection) {
-            showSourceActorSelection = false;
-            return;
-        }
-        availableSourceActors = game.actors
-            .filter(actor => actor.isOwner)
-            .filter(actor => actor.type === "character" || actor.type === "npc")
-            .filter(actor => actor.id !== targetActorDetails.id)
-            .filter(actor => actor.id !== sourceActorDetails.id)
-            .map(actor => new DefaultActorDetails({
-                id: actor.id,
-                name: actor.name,
-                avatarUrl: actor.img,
-                initials: DefaultActorDetails.getInitialsFromName(actor.name)
-            }));
-        showSourceActorSelection = true;
-    }
-
     async function prepareRecipes() {
         craftingAssessments = await fabricateAPI.crafting.summariseRecipes({
             targetActorId: targetActorDetails.id,
@@ -206,80 +129,23 @@
         salvageProcess = new DefaultSalvageProcess({ componentName: salvageAssessment.componentName });
     }
 
-    onMount(async () => {
-        await loadActorDetails();
+    async function load() {
         await prepareRecipes();
         await prepareComponents();
-    });
+    }
+
+    async function handleSourceActorChanged(event: CustomEvent<ActorDetails>) {
+        sourceActorDetails = event.detail;
+        await load();
+    }
+
+    onMount(load);
 
 </script>
 
 <div class="flex flex-col w-full h-full">
     <div class="flex flex-row">
-        <AppBar gridColumns="grid-cols-3" slotLead="place-self-start" slotTrail="w-full h-full relative" background="bg-surface-600" class="w-full">
-            <svelte:fragment slot="lead">
-                <div class="space-x-4 flex place-items-center">
-                    <Avatar src="{targetActorDetails.avatarUrl}" initials="{targetActorDetails.initials}" width="w-16" rounded="rounded-full" />
-                    <h1 class="mb-0 text-xl">{targetActorDetails.name}</h1>
-                </div>
-            </svelte:fragment>
-            <svelte:fragment slot="trail">
-                {#if sourceActorDetails.id === targetActorDetails.id}
-                    <div class="flex w-full justify-center">
-                        <a class="btn variant-filled-primary text-sm text-black" on:click={toggleSourceActorSelectionMenu}>
-                            <span><i class="fa-solid fa-box-open"></i></span>
-                            <span>Craft from another source</span>
-                        </a>
-                    </div>
-                {:else if availableSourceActors.length > 0}
-                    <div class="flex w-full justify-start">
-                        <div class="space-x-4 place-items-center cursor-pointer inline-flex" on:auxclick={clearSourceActor} on:click={toggleSourceActorSelectionMenu}>
-                            <div class="relative">
-                                <span class="text-black text-lg badge-icon variant-filled-tertiary absolute w-7 h-7 -top-0 -right-0 z-10">
-                                    <i class="fa-solid fa-box-open"></i>
-                                </span>
-                                <Avatar src="{sourceActorDetails.avatarUrl}" initials="{sourceActorDetails.initials}" width="w-16" rounded="rounded-full" class="no-img-border" />
-                            </div>
-                            <h2 class="text-xl">{sourceActorDetails.name}</h2>
-                        </div>
-                    </div>
-                {/if}
-                {#if showSourceActorSelection}
-                    <div class="absolute rounded-lg bg-surface-700 top-14 -left-3 w-full border-primary-300 border space-x-0 z-10"
-                         use:clickOutside
-                         on:clickOutside={() => { showSourceActorSelection = false; }}>
-                        <div class="p-4">
-                            <div class="input-group input-group-divider grid-cols-[auto_1fr_auto]">
-                                <div class="input-group-shim"><i class="fa-solid fa-magnifying-glass"></i></div>
-                                <input class="input h-full rounded-none p-2 text-black placeholder-gray-500" type="search" placeholder="Actor name..." bind:value={sourceActorName} />
-                            </div>
-                        </div>
-                        <div class="scroll scroll-secondary overflow-x-hidden overflow-y-auto max-h-48 snap-y snap-mandatory scroll-smooth scroll-px-4">
-                            <ListBox>
-                                {#each filteredSourceActors as availableSourceActor}
-                                    <ListBoxItem bind:group={sourceActorDetails}
-                                                 class="snap-start"
-                                                 name="Source Actor"
-                                                 hover="hover:bg-primary-400 hover:text-black"
-                                                 rounded="rounded-none"
-                                                 value="{availableSourceActor}"
-                                                 on:click={() => {setSourceActor(availableSourceActor);}}>
-                                        <svelte:fragment slot="lead">
-                                            <Avatar src="{availableSourceActor.avatarUrl}"
-                                                    initials="{availableSourceActor.initials}"
-                                                    width="w-8"
-                                                    rounded="rounded-full"
-                                                    class="inline-flex" />
-                                        </svelte:fragment>
-                                        <span>{availableSourceActor.name}</span>
-                                    </ListBoxItem>
-                                {/each}
-                            </ListBox>
-                        </div>
-                    </div>
-                {/if}
-            </svelte:fragment>
-        </AppBar>
+        <ActorCraftingAppHeader targetActorDetails={targetActorDetails} sourceActorDetails={sourceActorDetails} on:sourceActorChanged={handleSourceActorChanged} />
     </div>
     {#if craftingProcess.isReady}
         <AppBar background="bg-surface-700 text-white">
