+++ src/applications/actorCraftingApp/ActorCraftingAppManager.ts
@@ -1,0 +1,3 @@
+const key = Symbol();
+
+export { key };
